# backend/openapi/saxlem-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/request-otp": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Request an Iraqi mobile OTP challenge",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Verify OTP and create a patient session",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Authenticate clinic staff with email and password",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Rotate an opaque refresh token",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Revoke one refresh session",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Revoke every session for the identity",
        "tags": [
          "authentication"
        ]
      }
    },
...
    "/api/v1/patients/me": {
      "get": {
        "operationId": "PatientsController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Patient account projection."
          },
          "400": {
            "description": "Request validation failed."
          },
          "401": {
            "description": "Bearer token is missing, invalid, or expired."
          },
          "403": {
            "description": "The authenticated identity is not an eligible patient."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the authenticated patient account and current patient",
        "tags": [
          "patients"
        ]
      }
    },
    "/api/v1/patients/profiles": {
      "get": {
        "operationId": "PatientsController_profiles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientProfileResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed."
          },
          "401": {
            "description": "Bearer token is missing, invalid, or expired."
          },
          "403": {
            "description": "The authenticated identity is not an eligible patient."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List patient profiles owned by the authenticated account",
        "tags": [
          "patients"
        ]
      },
      "post": {
        "operationId": "PatientsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientProfileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed."
          },
          "401": {
            "description": "Bearer token is missing, invalid, or expired."
          },
          "403": {
            "description": "The authenticated identity is not an eligible patient."
          },
          "409": {
            "description": "Self-profile invariant would be violated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a self or family-member patient profile",
        "tags": [
          "patients"
        ]
      }
    },
    "/api/v1/patients/profiles/{id}": {
      "get": {
        "operationId": "PatientsController_profile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientProfileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed."
          },
          "401": {
            "description": "Bearer token is missing, invalid, or expired."
          },
          "403": {
            "description": "The authenticated identity is not an eligible patient."
          },
          "404": {
            "description": "Profile does not exist in this patient account."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one owned patient profile",
        "tags": [
          "patients"
        ]
      },
      "patch": {
        "operationId": "PatientsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientProfileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed."
          },
          "401": {
            "description": "Bearer token is missing, invalid, or expired."
          },
          "403": {
            "description": "The authenticated identity is not an eligible patient."
          },
          "409": {
            "description": "Profile is archived or its version is stale."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an owned active profile using optimistic concurrency",
        "tags": [
          "patients"
        ]
      },
      "delete": {
        "operationId": "PatientsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchivePatientProfileDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Request validation failed."
          },
          "401": {
            "description": "Bearer token is missing, invalid, or expired."
          },
          "403": {
            "description": "The authenticated identity is not an eligible patient."
          },
          "409": {
            "description": "Profile is active, archived, or its version is stale."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive a family-member profile; no physical deletion occurs",
        "tags": [
          "patients"
        ]
      }
    },
    "/api/v1/patients/active": {
      "post": {
        "operationId": "PatientsController_activate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivatePatientProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated patient account projection."
          },
          "400": {
            "description": "Request validation failed."
          },
          "401": {
            "description": "Bearer token is missing, invalid, or expired."
          },
          "403": {
            "description": "The authenticated identity is not an eligible patient."
          },
          "409": {
            "description": "Archived profiles cannot become active."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Choose the current active patient profile",
        "tags": [
          "patients"
        ]
      }
    }
  },
  "info": {
    "title": "Saxlem API",
    "description": "Versioned Saxlem backend contract",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RequestOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+9647500000000"
          }
        },
        "required": [
          "phone"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string"
          },
          "platform": {
            "type": "string",
            "enum": [
              "android",
              "ios",
              "web",
              "desktop"
            ]
          },
          "userAgent": {
            "type": "string"
          },
          "challengeId": {
            "type": "string"
          },
          "otp": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "deviceId",
          "platform",
          "challengeId",
          "otp"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string"
          },
          "platform": {
            "type": "string",
            "enum": [
              "android",
              "ios",
              "web",
              "desktop"
            ]
          },
          "userAgent": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "deviceId",
          "platform",
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string"
          },
          "platform": {
            "type": "string",
            "enum": [
              "android",
              "ios",
              "web",
              "desktop"
            ]
          },
          "userAgent": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "deviceId",
          "platform",
          "refreshToken"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "PatientProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "enum": [
              "female",
              "male",
              "unspecified"
            ]
          },
          "relationship": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "version": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "dateOfBirth",
          "gender",
          "relationship",
          "active",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreatePatientProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Shilan"
          },
          "lastName": {
            "type": "string",
            "example": "Ahmed"
          },
          "dateOfBirth": {
            "type": "string",
            "example": "1995-04-21",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "enum": [
              "female",
              "male",
              "unspecified"
            ]
          },
          "relationship": {
            "type": "string",
            "enum": [
              "me",
              "mother",
              "father",
              "son",
              "daughter",
              "brother",
              "sister",
              "grandfather",
              "grandmother",
              "wife",
              "husband",
              "other"
            ]
          }
        },
        "required": [
          "firstName",
          "lastName",
          "dateOfBirth",
          "gender",
          "relationship"
        ]
      },
      "UpdatePatientProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Shilan"
          },
          "lastName": {
            "type": "string",
            "example": "Ahmed"
          },
          "dateOfBirth": {
            "type": "string",
            "example": "1995-04-21",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "enum": [
              "female",
              "male",
              "unspecified"
            ]
          },
          "version": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "firstName",
          "lastName",
          "dateOfBirth",
          "gender",
          "version"
        ]
      },
      "ActivatePatientProfileDto": {
        "type": "object",
        "properties": {
          "profileId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "profileId"
        ]
      },
      "ArchivePatientProfileDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "version"
        ]
      }
    }
  }
}